Add support for null query browsing and additional background filters
Adds support for a default query (q) to send Solr when none is specified.  This allows straight user browsing with starting query, depending of course on being configured with a suitable catch-all starting point query.  Two "match-everything" examples are included.

Additionally adds support for configurable behind-the-scenes filters (fq), which may be used to restrict Solr queries arbitrarily.  This is similar to what's accomplished with the namespace restriction, except it is open-ended.